requirements.txt
asgiref==3.6.0
attrs==22.2.0
boto3==1.20.54
botocore==1.23.54
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.2.0
Django==4.1.5
django-auditlog==2.2.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-heroku==0.3.1
django-import-export==3.1.0
django-ipware==5.0.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
drf-yasg==1.21.5
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
importlib-resources==5.10.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==4.17.3
MarkupPy==1.14
MarkupSafe==2.1.2
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
psycopg2==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.3.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.13.0
mailersend==0.5.6
django-rest-passwordreset==1.3.0
redis==5.0.1
mysqlclient==2.2.1
pytest-django==4.8.0
pytest-cov==4.1.0
drf-api-logger==1.1.15
django-csp==3.8
cryptography==42.0.5
